pyproject.toml
[tool.poetry]
name = "kafka-app-template"
version = "0.1.0"
description = ""
authors = ["limdongjin <geniuslim27@gmail.com>"]
readme = "README.md"
packages = [{include = "kafka_app_template"}]

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^2.0.2"
protobuf = "^3.20.3"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
